+++ ll/tests/reg_format.rs
@@ -1,16 +1,17 @@
-use w5500_ll::SocketInterrupt;
+use w5500_ll::{Mode, PhyCfg, SocketInterrupt};
 
 #[test]
 fn sn_ir_none() {
     let snir = SocketInterrupt::default();
     assert_eq!(
-        format!("{}", snir),
-        r#"Socket interrupts:
-CON: not raised
-DISCON: not raised
-RECV: not raised
-SENDOK: not raised
-"#
+        format!("{:#}", snir),
+        r#"SocketInterrupt {
+    con_raised: false,
+    discon_raised: false,
+    recv_raised: false,
+    timeout_raised: false,
+    sendok_raised: false,
+}"#
     );
 }
 
@@ -19,13 +20,14 @@ fn sn_ir_all() {
     let snir: SocketInterrupt = u8::MAX.into();
 
     assert_eq!(
-        format!("{}", snir),
-        r#"Socket interrupts:
-CON: raised
-DISCON: raised
-RECV: raised
-SENDOK: raised
-"#
+        format!("{:#}", snir),
+        r#"SocketInterrupt {
+    con_raised: true,
+    discon_raised: true,
+    recv_raised: true,
+    timeout_raised: true,
+    sendok_raised: true,
+}"#
     );
 }
 
@@ -34,12 +36,45 @@ fn sn_ir_partial() {
     let snir: SocketInterrupt = SocketInterrupt::DISCON_MASK.into();
 
     assert_eq!(
-        format!("{}", snir),
-        r#"Socket interrupts:
-CON: not raised
-DISCON: raised
-RECV: not raised
-SENDOK: not raised
-"#
+        format!("{:#}", snir),
+        r#"SocketInterrupt {
+    con_raised: false,
+    discon_raised: true,
+    recv_raised: false,
+    timeout_raised: false,
+    sendok_raised: false,
+}"#
     );
+}
+
+#[test]
+fn mode() {
+    let mode: Mode = Mode::default();
+
+    assert_eq!(
+        format!("{:#}", mode),
+        r#"Mode {
+    wol_enabled: false,
+    pb_enabled: false,
+    pppoe_enabled: false,
+    farp_enabled: false,
+}"#,
+    )
+}
+
+#[test]
+fn phy_cfg() {
+    let phy_cfg: PhyCfg = PhyCfg::default();
+
+    assert_eq!(
+        format!("{:#}", phy_cfg),
+        r#"PhyCfg {
+    opmd: Ok(
+        Auto,
+    ),
+    dpx: Half,
+    spd: Mbps10,
+    lnk: Down,
+}"#,
+    )
 }