+++ ll/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `Display` for `SocketInterruptMask`.
+- Added `Display` for `SocketMode`.
+- Added `Display` for `Mode`.
+
+### Changed
+- Changed the display formatting for `LinkStatus` and `SocketInterrupt`.
+
 ## [0.7.0] - 2021-02-23
 ### Changed
 - Updated `defmt` dependency from 0.1.3 to 0.2.0.